UI: mark visually motor positions outside of limits
Changes behavior of motor widgets when position outside of motor
limits is entered.

If entered position is outside of limits, change the widget style to
'invalid' mode. Block sending the 'invalid' position to the backend.

When focus is lost, restore widget to display motor's real position.

This changes the current behavior, where it's possible to enter
and send a position outside of limits. The motor can refuse to move
to that position, but the widget will still display the invalid value.

Thus the displayed value in the UI will be different from the motor's
real position, without any visual indications.