Adds tooltip for field description in generic form
This also introduces refactor. The `customFieldTemplate` has been moved
to be a separate functional component.
This should make the `GenericTaskForm` component a bit thinner, while
also making a progress in the `class->functional` rewrite of the
components.

It also introduces a slight change in the global css.
By default bootstrap provides a higher z-index value for modals than for
tooltips. It is problematic for this use-case of describing form fields
in a modal.
So I've increased z-index for tooltips that are rendered by rjsf.

There was also a slight change. There is no conditional rendering for
`errors` property in `CustomFieldTemplate` - it is already a React
element - it's always truthy.