SelectProposal: Add a way to show/hide tabs
At MAX IV we don't distinguish between different types of sessions
(scheduled, unscheduled, other beamline session).
Our sessions are created "on the fly" and are not tied to a particular
beam time, thus it's not necessary to split them.

This commit makes it possible to opt-out of displaying some of the tabs
via `ui.yaml` config file.