Improve Python docstrings
The intention of this commit is to improve the Python docstrings
so that tools like Sphinx can generate a more useful output.

For this purpose, the Ruff `D` rules are enabled except `D1`,
which means all the rules from the `pydocstyle` group
except the rules about missing docstrings.

This does not attempt to unify type hints between
docstrings and annotations.
This does not attempt to add missing docstrings,
or fix incomplete or incorrect docstrings.

Most of this was done automatically by Ruff with `ruff check --fix`.
Then some manual editing was done to fix remaining issues:

* Some light rephrasing, in particular to extract a title line.
* Some formatting of code embedded in docstrings.
* Some typo fixes here and there.

Still, a lot of docstrings do not follow the Google style,
although that is our chosen style according to our style guidelines.